backend: extract model catalog endpoints to routers/models.py; include router and remove handlers from main.py (closes #12)
